# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Open API - E Commerce",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/cart": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Create new Product in cart",
                "description": "Create new Product in cart",
                "operationId": "ba73c99e1f1e984d670c8f11ee9353dd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "customer_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "quantity": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "product_id": {
                                            "type": "integer",
                                            "example": "2"
                                        },
                                        "customer_id": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/cart/product/{Id}": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Find Product in cart by ID",
                "description": "Returns a single product",
                "operationId": "34b82b5d7a493d5d9d91b8665c73016b",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "Id Cart",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Samsung S24"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "samsung memiliki ....."
                                        },
                                        "price": {
                                            "type": "integer",
                                            "example": "34000000"
                                        },
                                        "stock": {
                                            "type": "integer",
                                            "example": "8"
                                        },
                                        "rate": {
                                            "type": "integer",
                                            "example": "5"
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "category": {
                                            "type": "integer",
                                            "example": "Samsung"
                                        },
                                        "quantity": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/api/cart/{Id}": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Show Product in Cart by Customer",
                "description": "Show Product in Cart by Customer",
                "operationId": "61cb087a0c8383ea1013678744b45198",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "Id Customer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "data",
                                    "type": "array",
                                    "items": {
                                        "title": "data[0]",
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "1"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Samsung S24"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "samsung memiliki ....."
                                            },
                                            "price": {
                                                "type": "integer",
                                                "example": "34000000"
                                            },
                                            "stock": {
                                                "type": "integer",
                                                "example": "8"
                                            },
                                            "rate": {
                                                "type": "integer",
                                                "example": "5"
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "category": {
                                                "type": "string",
                                                "example": "Samsung"
                                            },
                                            "quantity": {
                                                "type": "integer",
                                                "example": "1"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "example": [
                                    {
                                        "id": "1",
                                        "name": "Samsung S24",
                                        "description": "Samsung S24 memiliki...",
                                        "price": "3000000",
                                        "stock": "10",
                                        "rate": "4",
                                        "category_id": "1",
                                        "category": "Samsung",
                                        "quantity": "2"
                                    },
                                    {
                                        "id": "2",
                                        "name": "Samsung S23",
                                        "description": "Samsung S23 memiliki...",
                                        "price": "2800000",
                                        "stock": "10",
                                        "rate": "4",
                                        "category_id": "1",
                                        "category": "Samsung",
                                        "quantity": "2"
                                    }
                                ]
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Delete Product in Cart by ID",
                "description": "Delete a single product",
                "operationId": "28c933a83894161134694bf0550db441",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "Id Cart",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    },
...
}